# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Which Docs
# during dev, should point to your local machine
host: localhost:5068
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/updateStatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: users_ctrl
    post:
      tags: ['User Module']
      description: Sets User Profile
      summary: Performs updation
      # used as the method name of the controller
      operationId: updateStatus
      parameters:
        - name: body
          in: body
          description: Update User Profile Info
          schema:
           $ref: '#/definitions/updateStatus'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/addSpeciality:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['speciality Module']
      description: contains all specility doctors are specialised in
      summary: deals with speciality
      # used as the method name of the controller
      operationId: addSpeciality
      parameters:
        - name: body
          in: body
          description: sets speciality name
          schema:
           $ref: '#/definitions/addSpeciality'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/addSpecialityService:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['speciality Module']
      description: contains all specility doctors are specialised in
      summary: deals with speciality
      # used as the method name of the controller
      operationId: addSpecialityService
      parameters:
        - name: body
          in: body
          description: sets speciality name
          schema:
           $ref: '#/definitions/addSpecialityService'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getSpeciality:
    # Doctors APIs
    x-swagger-router-controller: speciality_Ctrl
    get:
      tags: ['speciality Module']
      description: Get all the specialities
      summary: available
      # used as the method name of the controller
      operationId: getSpeciality
      parameters:
        - name: id
          in: query
          description: get all specialities
          required: false
          type: string        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/updatePatientNotificationStatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['speciality Module']
      description: Sets patient notification acitvated/deactivated for Speciality
      summary: Performs updation
      # used as the method name of the controller
      operationId: updatePatientNotificationStatus
      parameters:
        - name: body
          in: body
          description: Update Speciality notification Info
          schema:
           $ref: '#/definitions/updatePatientNotificationStatus'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
...
  /api/getSpecialityList:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['specialities Module']
      description: Get Specialty List
      summary: Get Specialty List
      # used as the method name of the controller
      operationId: getSpecialityList
      parameters:
        - name: body
          in: body
          description: Get Speciality List
          schema:
           $ref: '#/definitions/getSpecialityList'
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  # /api/getReferredBy:
  #   # binds a127 app logic to a route
  #   x-swagger-router-controller: doctor_ctrl
  #   post:
  #     tags: ['specialities Module']
  #     description: Get Specialty List
  #     summary: Get Specialty List
  #     # used as the method name of the controller
  #     operationId: getReferredBy
  #     parameters:
  #       - name: name
  #         in: body
  #         description: Get getReferred By
  #         schema:
  #          $ref: '#/definitions/getReferredBy'
  #         required: true
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           # a pointer to a definition
  #           $ref: "#/definitions/appResponse"
  #       # responses may fall through to errors
  #       default:
  #         description: Error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  # /api/getReferredTo:
  #   # binds a127 app logic to a route
  #   x-swagger-router-controller: doctor_ctrl
  #   post:
  #     tags: ['specialities Module']
  #     description: Get Specialty List
  #     summary: Get Specialty List
  #     # used as the method name of the controller
  #     operationId: getReferredTo
  #     parameters:
  #       - name: name
  #         in: body
  #         description: Get getReferred To
  #         schema:
  #          $ref: '#/definitions/getReferredTo'
  #         required: true
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           # a pointer to a definition
  #           $ref: "#/definitions/appResponse"
  #       # responses may fall through to errors
  #       default:
  #         description: Error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  /api/getSpecialityById/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    get:
      tags: ['Speciality Module']
      description: Get Speciality By Id
      summary: Get Speciality By Id
      # used as the method name of the controller
      operationId: getSpecialityById
      parameters:
        - name: id
          type: string
          in: path
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
  /api/updateSpecialityByAdmin:
    # Doctors APIs
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['Speciality Module']
      description: Update Speciality Information within the network
      summary: Update Speciality Information within the network
      # used as the method name of the controller
      operationId: updateSpecialityByAdmin
      parameters:
        - name: name
          in: body
          description: Update Speciality Information within the network
          schema:
            $ref: '#/definitions/updateSpecialityByAdmin'
          required: true
          # type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deleteSpeciality:
    # Doctors APIs
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['Speciality Module']
      description: Update Speciality Information within the network
      summary: Update Speciality Information within the network
      # used as the method name of the controller
      operationId: deleteSpeciality
      parameters:
        - name: name
          in: body
          description: Update Speciality Information within the network
          schema:
            $ref: '#/definitions/deleteSpeciality'
          required: true
          # type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/specialityServiceDelete:
    # Doctors APIs
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['Speciality Module']
      description: Update Speciality Information within the network
      summary: Update Speciality Information within the network
      # used as the method name of the controller
      operationId: specialityServiceDelete
      parameters:
        - name: name
          in: body
          description: Update Speciality Information within the network
          schema:
            $ref: '#/definitions/specialityServiceDelete'
          required: true
          # type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/updateSpecialityStatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['Speciality Module']
      description: Sets Speciality Profile acitvated/deactivated
      summary: Performs updation
      # used as the method name of the controller
      operationId: updateSpecialityStatus
      parameters:
        - name: body
          in: body
          description: Update Speciality Profile Info
          schema:
           $ref: '#/definitions/updateSpecialityStatus'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/specialityServiceStatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['Speciality Module']
      description: Sets Speciality Profile acitvated/deactivated
      summary: Performs updation
      # used as the method name of the controller
      operationId: specialityServiceStatus
      parameters:
        - name: body
          in: body
          description: Update Speciality Profile Info
          schema:
           $ref: '#/definitions/specialityServiceStatus'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getSpecialityServiceList:
    # binds a127 app logic to a route
    x-swagger-router-controller: speciality_Ctrl
    post:
      tags: ['Speciality Module']
      description: Get Service to the Speciality
      summary: Get Service to the Speciality
      # used as the method name of the controller
      operationId: getSpecialityServiceList
      parameters:
        - name: body
          in: body
          description: Get Service to the Speciality
          schema:
           $ref: '#/definitions/getSpecialityServiceList'
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  deleteSpeciality:
    type: object
    properties:
      id:
        type: string
  specialityServiceDelete:
    type: object
    properties:
      id:
        type: string
...
  updateStatus:
    type: object
    properties:
      id:
        type: string
      status:
        type: string
...
  updatePatientNotificationStatus:
    type: object
    properties:
      id:
        type: string
      status:
        type: string        
...
  addSpeciality: 
    type: object
    properties:
      specialityCode:
        type: string
      specialityName:
        type: string
      specialityDescp:
        type: string
...
  addSpecialityService: 
    type: object
    properties:
      serviceCode:
        type: string
      serviceName:
        type: string
      serviceDescp:
        type: string
...
  getSpecialityList:
    type: object
    properties:
      page:
        type: string
      count:
        type: string
...
  updateSpecialityByAdmin: 
    type: object
    properties:
      specialityCode:
        type: string
      specialityName:
        type: string
...
  updateSpecialityStatus:
    type: object
    properties:
      id:
        type: string
      status:
        type: string
  specialityServiceStatus:
    type: object
    properties:
      id:
        type: string
      status:
        type: string
...
  getSpecialityServiceList:
    type: object
    properties:
      page:
        type: string
      count:
        type: string                                
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...